connecting posts and users adding auth checks set up user status

+++ 18-1-rest-api/controllers/feed.js
@@ -4,6 +4,7 @@ const path = require('path');
 const { validationResult } = require('express-validator');
 
 const Post = require('../models/post');
+const User = require('../models/user');
 
 exports.getPosts = (req, res, next) => {
 	const currentPage = req.query.page || 1;
@@ -20,13 +21,11 @@ exports.getPosts = (req, res, next) => {
 				.limit(perPage);
 		})
 		.then((posts) => {
-			res
-				.status(200)
-				.json({
-					message: 'Fetched posts successfully',
-					posts: posts,
-					totalItems: totalItems,
-				});
+			res.status(200).json({
+				message: 'Fetched posts successfully',
+				posts: posts,
+				totalItems: totalItems,
+			});
 		})
 		.catch((err) => {
 			if (!err.statusCode) {
@@ -56,20 +55,30 @@ exports.postPost = (req, res, next) => {
 	const title = req.body.title;
 	const content = req.body.content;
 	const imageUrl = req.file.path.replace('\\', '/');
+	let creator;
 
 	const post = new Post({
 		title: title,
 		content: content,
 		imageUrl: imageUrl,
-		creator: { name: 'Gary' },
+		creator: req.userId,
 	});
 
 	post
 		.save()
-		.then((result) => {
+		.then(() => {
+			return User.findById(req.userId);
+		})
+		.then((user) => {
+			creator = user;
+			user.posts.push(post);
+			return user.save();
+		})
+		.then(() => {
 			res.status(201).json({
 				message: 'Post created succesfully',
-				post: result,
+				post: post,
+				creator: { _id: creator._id, name: creator.name },
 			});
 		})
 		.catch((err) => {
@@ -136,6 +145,12 @@ exports.updatePost = (req, res, next) => {
 				throw error; //throws to catch
 			}
 
+			if (post.creator.toString() !== req.userId) {
+				const error = new Error('Not authorized');
+				error.statusCode = 403;
+				throw error;
+			}
+
 			if (imageUrl !== post.imageUrl) {
 				clearImage(post.imageUrl);
 			}
@@ -168,11 +183,22 @@ exports.deletePost = (req, res, next) => {
 				throw error; //throws to catch
 			}
 
-			// check logged in  user
+			if (post.creator.toString() !== req.userId) {
+				const error = new Error('Not authorized');
+				error.statusCode = 403;
+				throw error;
+			}
 
 			clearImage(post.imageUrl);
 
 			return Post.findByIdAndRemove(postId);
+		})
+		.then(() => {
+			return User.findById(req.userId);
+		})
+		.then((user) => {
+			user.posts.pull(postId);
+			return user.save();
 		})
 		.then(() => {
 			res.status(200).json({ message: 'Deleted post.' });
